Add optional logging
So that users can more easily resume harvesting if there are any network
issues, add logging of network requests and resumption tokens.

Note that, by default, a null logger is used which writes to the
appropriate null device for the platform. As File::NULL was only added
in Ruby 1.9.3, we backport this logic based on the implementation from
the backports gem.