Add explicit argument parsing & default to oai_dc
Rather than users having to know the string query parameters supported
by the OAI-PMH protocol, specify our own symbol arguments and handle
various value types (e.g. accepting Dates and Times as well as Strings).

As every OAI-PMH repository has to provide Dublin Core as a metadata
format, default to "oai_dc" as the metadata prefix for all requests that
require it. It can still be overridden by the user but it is a sensible
default for most repositories.