Switch Identify parser to SAX parsing
Instead of parsing the full response only to use some of it to populate
value objects, use a SAX parser to only use parts of the XML we care
about. This means we can gather parts we want to return directly to the
user (e.g. descriptions) in a raw string.

As this means that responses are built up incrementally, the Identify
class is now largely an empty shell with accessors for each field. We
could change this to be slightly less mutable by building them in one go
but this seemed easier for now.