Extract Parsers & document every public method
Rather than having the Paginator both handle flow control and parsing of
XML, extract dedicated response parsers which handle converting a
HTTP response body into an Enumerator of items.

This trades performance off maintainability as there are more
intermediate objects between user and their resulting value objects but
the separation of concerns might be worth it.

In addition to the extensive API documentation in the README, also
document every public method so that this information is included on
sites such as RubyDoc.info, etc. which automatically parse comments from
gems.