Return a compressed output for an element or its metdata
When extracting the original xml for either the element or its metadata,
the current implementation includes the default indentation of 2 spaces for the
heirarchy of the xml. Instead return the most compressed output possible
for the element/metadata, which then gives the user the ability to choose
how to format the xml.