Add timeout option to the Repository
GitHub: https://github.com/altmetric/fieldhand/issues/9

Allow users to set an HTTP timeout in seconds which will be used by the
underlying HTTP client. It currently defaults to 60 seconds to match the
defaults from Ruby 2.3 onwards.

Note this retains API compatibility with previous versions of Fieldhand
by allowing users to pass a Logger directly or by passing the new style
of options as a Hash.