Move to dedicated Fieldhand organisation
Given the origins of this library, I would like to retain some ownership
and maintenance responsibilities despite leaving Altmetric. To keep
things neutral, move the source code to its own GitHub organisation.

This does not change the current owners of the published RubyGem but
makes the ownership and licensing clearer.